refactor: ItemGroup을 OrderItemGroup으로 이름 변경
